+++ rustywind-core/CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ## [Unreleased]
 
+### Performance
+
+- In MacOS limit number of threads to 4 for up to a 400% performance boost
+
 ### Refactor
 
 - Completely refactored the public API, now all the functionality is in the `RustyWind` struct
+- Remove `rayon` and use a custom scoped thread implementation
 
 ### Changed